+++ azurerm/helpers/azure/servicebus.go
@@ -1,54 +1,13 @@
 package azure
 
 import (
-	"fmt"
-	"log"
-
-	"github.com/Azure/azure-sdk-for-go/services/servicebus/mgmt/2017-04-01/servicebus"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
 
-func ExpandServiceBusAuthorizationRuleRights(d *schema.ResourceData) *[]servicebus.AccessRights {
-	rights := make([]servicebus.AccessRights, 0)
-
-	if d.Get("listen").(bool) {
-		rights = append(rights, servicebus.Listen)
-	}
-
-	if d.Get("send").(bool) {
-		rights = append(rights, servicebus.SendEnumValue)
-	}
-
-	if d.Get("manage").(bool) {
-		rights = append(rights, servicebus.Manage)
-	}
-
-	return &rights
-}
-
-func FlattenServiceBusAuthorizationRuleRights(rights *[]servicebus.AccessRights) (listen, send, manage bool) {
-	// zero (initial) value for a bool in go is false
-
-	if rights != nil {
-		for _, right := range *rights {
-			switch right {
-			case servicebus.Listen:
-				listen = true
-			case servicebus.SendEnumValue:
-				send = true
-			case servicebus.Manage:
-				manage = true
-			default:
-				log.Printf("[DEBUG] Unknown Authorization Rule Right '%s'", right)
-			}
-		}
-	}
-
-	return listen, send, manage
-}
-
 // shared schema
 func MergeSchema(a map[string]*schema.Schema, b map[string]*schema.Schema) map[string]*schema.Schema {
+	// TODO: Deprecate and remove this
+
 	s := map[string]*schema.Schema{}
 
 	for k, v := range a {
@@ -60,67 +19,4 @@ func MergeSchema(a map[string]*schema.Schema, b map[string]*schema.Schema) map[s
 	}
 
 	return s
-}
-
-func ServiceBusAuthorizationRuleSchemaFrom(s map[string]*schema.Schema) map[string]*schema.Schema {
-	authSchema := map[string]*schema.Schema{
-		"listen": {
-			Type:     schema.TypeBool,
-			Optional: true,
-			Default:  false,
-		},
-
-		"send": {
-			Type:     schema.TypeBool,
-			Optional: true,
-			Default:  false,
-		},
-
-		"manage": {
-			Type:     schema.TypeBool,
-			Optional: true,
-			Default:  false,
-		},
-
-		"primary_key": {
-			Type:      schema.TypeString,
-			Computed:  true,
-			Sensitive: true,
-		},
-
-		"primary_connection_string": {
-			Type:      schema.TypeString,
-			Computed:  true,
-			Sensitive: true,
-		},
-
-		"secondary_key": {
-			Type:      schema.TypeString,
-			Computed:  true,
-			Sensitive: true,
-		},
-
-		"secondary_connection_string": {
-			Type:      schema.TypeString,
-			Computed:  true,
-			Sensitive: true,
-		},
-	}
-	return MergeSchema(s, authSchema)
-}
-
-func ServiceBusAuthorizationRuleCustomizeDiff(d *schema.ResourceDiff, _ interface{}) error {
-	listen, hasListen := d.GetOk("listen")
-	send, hasSend := d.GetOk("send")
-	manage, hasManage := d.GetOk("manage")
-
-	if !hasListen && !hasSend && !hasManage {
-		return fmt.Errorf("One of the `listen`, `send` or `manage` properties needs to be set")
-	}
-
-	if manage.(bool) && !listen.(bool) && !send.(bool) {
-		return fmt.Errorf("if `manage` is set both `listen` and `send` must be set to true too")
-	}
-
-	return nil
 }
