+++ azurerm/internal/services/servicebus/servicebus_namespace_authorization_rule_resource.go
@@ -34,7 +34,7 @@ func resourceArmServiceBusNamespaceAuthorizationRule() *schema.Resource {
 		},
 
 		// function takes a schema map and adds the authorization rule properties to it
-		Schema: azure.ServiceBusAuthorizationRuleSchemaFrom(map[string]*schema.Schema{
+		Schema: ServiceBusAuthorizationRuleSchemaFrom(map[string]*schema.Schema{
 			"name": {
 				Type:         schema.TypeString,
 				Required:     true,
@@ -52,7 +52,7 @@ func resourceArmServiceBusNamespaceAuthorizationRule() *schema.Resource {
 			"resource_group_name": azure.SchemaResourceGroupName(),
 		}),
 
-		CustomizeDiff: azure.ServiceBusAuthorizationRuleCustomizeDiff,
+		CustomizeDiff: ServiceBusAuthorizationRuleCustomizeDiff,
 	}
 }
 
@@ -83,7 +83,7 @@ func resourceArmServiceBusNamespaceAuthorizationRuleCreateUpdate(d *schema.Resou
 	parameters := servicebus.SBAuthorizationRule{
 		Name: &name,
 		SBAuthorizationRuleProperties: &servicebus.SBAuthorizationRuleProperties{
-			Rights: azure.ExpandServiceBusAuthorizationRuleRights(d),
+			Rights: ExpandServiceBusAuthorizationRuleRights(d),
 		},
 	}
 
@@ -133,7 +133,7 @@ func resourceArmServiceBusNamespaceAuthorizationRuleRead(d *schema.ResourceData,
 	d.Set("resource_group_name", resGroup)
 
 	if properties := resp.SBAuthorizationRuleProperties; properties != nil {
-		listen, send, manage := azure.FlattenServiceBusAuthorizationRuleRights(properties.Rights)
+		listen, send, manage := FlattenServiceBusAuthorizationRuleRights(properties.Rights)
 		d.Set("manage", manage)
 		d.Set("listen", listen)
 		d.Set("send", send)
