+++ azurerm/internal/services/servicebus/servicebus_topic_authorization_rule_resource.go
@@ -33,7 +33,7 @@ func resourceArmServiceBusTopicAuthorizationRule() *schema.Resource {
 			Delete: schema.DefaultTimeout(30 * time.Minute),
 		},
 
-		Schema: azure.ServiceBusAuthorizationRuleSchemaFrom(map[string]*schema.Schema{
+		Schema: ServiceBusAuthorizationRuleSchemaFrom(map[string]*schema.Schema{
 			"name": {
 				Type:         schema.TypeString,
 				Required:     true,
@@ -58,7 +58,7 @@ func resourceArmServiceBusTopicAuthorizationRule() *schema.Resource {
 			"resource_group_name": azure.SchemaResourceGroupName(),
 		}),
 
-		CustomizeDiff: azure.ServiceBusAuthorizationRuleCustomizeDiff,
+		CustomizeDiff: ServiceBusAuthorizationRuleCustomizeDiff,
 	}
 }
 
@@ -89,7 +89,7 @@ func resourceArmServiceBusTopicAuthorizationRuleCreateUpdate(d *schema.ResourceD
 	parameters := servicebus.SBAuthorizationRule{
 		Name: &name,
 		SBAuthorizationRuleProperties: &servicebus.SBAuthorizationRuleProperties{
-			Rights: azure.ExpandServiceBusAuthorizationRuleRights(d),
+			Rights: ExpandServiceBusAuthorizationRuleRights(d),
 		},
 	}
 
@@ -141,7 +141,7 @@ func resourceArmServiceBusTopicAuthorizationRuleRead(d *schema.ResourceData, met
 	d.Set("resource_group_name", resourceGroup)
 
 	if properties := resp.SBAuthorizationRuleProperties; properties != nil {
-		listen, send, manage := azure.FlattenServiceBusAuthorizationRuleRights(properties.Rights)
+		listen, send, manage := FlattenServiceBusAuthorizationRuleRights(properties.Rights)
 		d.Set("listen", listen)
 		d.Set("send", send)
 		d.Set("manage", manage)
